[Refactor] * linter applied * task moved to its own .py file (to avoid circular import error)
